start computing a static list of JS files
- Renamed hasher.js to scanner.js as now it can return both the list of JS files
  in the subtree as well as the hash-code.
- This paves the way to us using the static list of files to validate the break
  point requests against rather than the dynamic list of files loaded in V8.
