+++ lib/uid.js
@@ -50,7 +50,7 @@ module.exports.get = function (baseDir, callback) {
       return;
     }
 
-    var hasher = require('./hasher.js');
-    hasher.compute(baseDir, callback);
+    var scanner = require('./scanner.js');
+    scanner.scan(baseDir, callback);
   });
 };
